% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_server.R
\name{ar.multivariate}
\alias{ar.multivariate}
\title{Multivariate version of AR}
\usage{
ar.multivariate(x, aic = TRUE, order.max = 1, na.action = na.fail,
  demean = TRUE, intercept = demean, type = "const", ...)
}
\description{
\code{ar.multivariate} returns arest model
}
