% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R, R/utils_server.R
\name{computeTP}
\alias{computeTP}
\title{Parent function for computing the MSE for a particular model}
\usage{
computeTP(nVar, time, error, gen_model, model1 = "ar", model2 = "var",
  phi, inno, K = 5, index_vars, lagNum = 1, error_metric = "mse")

computeTP(nVar, time, error, gen_model, model1 = "ar", model2 = "var",
  phi, inno, K = 5, index_vars, lagNum = 1, error_metric = "mse")
}
\arguments{
\item{model}{A fitted ts model, currently ar, arest (from ar.multivariate) and var are supported}

\item{model}{A fitted ts model, currently ar, arest (from ar.multivariate) and var are supported}
}
\value{
A matrix of residuals, with columns being the variables.

A matrix of residuals, with columns being the variables.
}
\description{
\code{extractResiduals} returns the residuals in matrix-format from a fitted timeseries model.

\code{extractResiduals} returns the residuals in matrix-format from a fitted timeseries model.
}
\seealso{
Other extractresiduals.ar: \code{\link{extractResiduals}}

Other extractresiduals.ar: \code{\link{extractResiduals}}
}
\concept{extractresiduals.ar}
