% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_server.R
\name{extractInno}
\alias{extractInno}
\title{Parent function for computing the residual covariance matrix for a particular model}
\usage{
extractInno(model)
}
\arguments{
\item{model}{A fitted ts model, currently ar, arest (from ar.multivariate) and var are supported}
}
\value{
A covariance matrix of the residuals.
}
\description{
\code{extractInno} returns the residual covariance matrix of a timeseries model.
}
\concept{extractInno.ar, extractInno.varest, extractInno.arest}
