% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_server.R
\name{extractResiduals}
\alias{extractResiduals}
\title{Parent function pointing to other functions that extract the residuals from a fitted timeseries model}
\usage{
extractResiduals(model)
}
\arguments{
\item{model}{A fitted ts model, currently ar, arest (from ar.multivariate) and var are supported}
}
\value{
A matrix of residuals, with columns being the variables.
}
\description{
\code{extractResiduals} returns the residuals in matrix-format from a fitted timeseries model.
}
\concept{extractresiduals.ar}
