% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_server.R
\name{computeTP}
\alias{computeTP}
\title{Parent function for computing the MSE for a particular model}
\usage{
computeTP(nVar, time, error, gen_model, model1 = "ar", model2 = "var",
  K = 5, index_vars, lagNum = 1, error_metric = "mse", model_params)
}
\arguments{
\item{model}{A fitted ts model, currently ar, arest (from ar.multivariate) and var are supported}
}
\value{
A matrix of residuals, with columns being the variables.
}
\description{
\code{computeTP} returns the residuals in matrix-format from a fitted timeseries model.
}
