% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{computeTP}
\alias{computeTP}
\title{Parent function for computing the MSE for a particular model}
\usage{
computeTP(nVar, time, error, model, phi, inno, K = 5)
}
\arguments{
\item{model}{A fitted ts model, currently ar, arest (from ar.multivariate) and var are supported}
}
\value{
A matrix of residuals, with columns being the variables.
}
\description{
\code{extractResiduals} returns the residuals in matrix-format from a fitted timeseries model.
}
\seealso{
Other extractresiduals.ar: \code{\link{extractResiduals}}
}
\concept{extractresiduals.ar}
